split index.js into mainly createClient.js and createServer.js also create a few other files containing common information and functions : * mcHexDigest.js which contain the function of the same name used in both createClient.js and createServer.js * ursa.js which handle the abstraction on top of ursa and ursa-purejs : used in both createClient.js and createServer.js * version.js used in index.js, createClient.js and createServer.js
