+++ src/version.js
@@ -1,0 +1,4 @@
+module.exports={
+  version:47,
+  minecraftVersion:'1.8.1'
+};
